chore: refac use hook on category
